chore: lib:watch uses a fixed tsconfig.json
